feat: Overhaul backend, use IoDriver trait across io_uring and generic
